Simplify thrift generation
