+++ lib/mix/tasks/elixir.ex
@@ -10,7 +10,7 @@ defmodule Mix.Tasks.Thrifter.Elixir do
     Mix.shell.info "\n--- Generating elixir client ---\n"
 
     clean_output_dir
-    generate_thrift_files
+    Thrift.generate(output: thrift_output_dir, language: "erl")
     generate_elixir_files
 
     Mix.shell.info "\Elixir client generated in #{Colors.green(client_dir)}\n"
@@ -18,6 +18,10 @@ defmodule Mix.Tasks.Thrifter.Elixir do
 
   defp client_dir do
     "gen/elixir-client"
+  end
+
+  def thrift_output_dir do
+    "#{client_dir}/src"
   end
 
   defp client_name do
@@ -33,9 +37,9 @@ defmodule Mix.Tasks.Thrifter.Elixir do
     options = [
       client_name: client_name,
       client_module_name: Macro.camelize(client_name),
-      service_name: service_name,
-      function_names: function_names,
-      structs: structs,
+      service_name: Thrift.Erlang.service_name(thrift_output_dir),
+      function_names: Thrift.Erlang.function_names(thrift_output_dir),
+      structs: Thrift.Erlang.structs(thrift_output_dir),
       version: Mix.Project.config[:version]
     ]
 
@@ -54,38 +58,6 @@ defmodule Mix.Tasks.Thrifter.Elixir do
 
       File.write! output, Templates.render(template, options)
     end
-  end
-
-  def generate_thrift_files do
-    Mix.shell.info "\nCompiling thrift client\n"
-
-    output = "#{client_dir}/src"
-    File.mkdir_p!(output)
-
-    Thrift.generate(output: output, language: "erl")
-  end
-
-  defp function_names do
-    "cd #{client_dir}/src; grep function_info *erl | grep \\(\\' | cut -d\\' -f2 | sort -u"
-    |> os_cmd
-    |> String.split
-    |> Enum.map(&String.to_atom/1)
-  end
-
-  defp service_name do
-    "cd #{client_dir}/src; grep function_info *erl | head -1 | cut -f1 -d."
-    |> os_cmd
-    |> String.strip(?\n)
-  end
-
-  defp structs do
-    "cd #{client_dir}/src; grep \"struct_info('\" *_types.erl | awk -F \"'\" '{ print $2 }'"
-    |> os_cmd
-    |> String.split
-  end
-
-  defp os_cmd(cmd) do
-    cmd |> to_char_list |> :os.cmd |> List.to_string
   end
 
 end
